refactor: Update attributeHandler.js to handle boolean attributes
The code changes in attributeHandler.js modify the logic to handle boolean attributes. A set of predefined boolean attributes is created, and if an attribute is present in the set and its value is truthy, the attribute is set on the element. If the value is falsy, the attribute is removed. This refactor improves the handling of boolean attributes in the codebase.