refactor: Update checkbox handling in element.advance.test.js
Refactor the code in element.advance.test.js to update the handling of checkboxes. Instead of using the 'checked' attribute with boolean values, the code now uses empty string ('') to represent a checked checkbox and the 'hidden' attribute to hide elements. This change improves consistency and compatibility with HTML standards.

Though, some problems has been detected, null and undefined as attribute makes it problematic